split appearance settings
