+++ iosApp/flare/Assets.xcassets/fa-newspaper.symbolset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "symbols" : [
+    {
+      "filename" : "fa-newspaper.svg",
+      "idiom" : "universal"
+    }
+  ]
+}
