+++ iosApp/flare/Assets.xcassets/fa-photo-film.symbolset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "symbols" : [
+    {
+      "filename" : "fa-photo-film.svg",
+      "idiom" : "universal"
+    }
+  ]
+}
